backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
torch==2.5.1
shap==0.46.0
google-genai==1.5.0
python-dotenv==1.0.1
pydantic==2.10.5
websockets==14.2
